Initial stab at configuring providers via Admin.
Changes to a custom Admin site to explicitly add the apps to the Admin
interface that we care about.
Uses django-constance for persisting settings to the database.
Allows the admin interface to store the provider config in the database
and read the setting from there on startup.
Shows only the configured providers to the user vs all available
providers.